Refactor function names

+++ Game/src/Scene/Pools/Asteroids/Asteroids.h
@@ -5,7 +5,7 @@
 class Asteroids : public Pool
 {
   public:
-    const size_t total = 3;
+    const int total = 3;
     const std::string file = "sphere.obj";
 
     virtual void Init(Scene &scene) override;
